// frontend/src/api/models/car-info-vm.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Resume Car Builder API
 * An API for Resume Car Builder
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface CarInfoVM
 */
export interface CarInfoVM {
    /**
     * 
     * @type {string}
     * @memberof CarInfoVM
     */
    id?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CarInfoVM
     */
    brand?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CarInfoVM
     */
    model?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CarInfoVM
     */
    yearOfManufacture?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CarInfoVM
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CarInfoVM
     */
    distance?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CarInfoVM
     */
    color?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CarInfoVM
     */
    price?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CarInfoVM
     */
    engine?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CarInfoVM
     */
    status?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CarInfoVM
     */
    carImageURL?: string | null;
}